UI: Move multiview render into a new class
