+++ UI/window-projector.hpp
@@ -2,6 +2,7 @@
 
 #include <obs.hpp>
 #include "qt-display.hpp"
+#include "multiview.hpp"
 
 enum class ProjectorType {
 	Source,
@@ -12,19 +13,6 @@ enum class ProjectorType {
 };
 
 class QMouseEvent;
-
-enum class MultiviewLayout : uint8_t {
-	HORIZONTAL_TOP_8_SCENES = 0,
-	HORIZONTAL_BOTTOM_8_SCENES = 1,
-	VERTICAL_LEFT_8_SCENES = 2,
-	VERTICAL_RIGHT_8_SCENES = 3,
-	HORIZONTAL_TOP_24_SCENES = 4,
-	HORIZONTAL_TOP_18_SCENES = 5,
-	SCENES_ONLY_4_SCENES = 6,
-	SCENES_ONLY_9_SCENES = 7,
-	SCENES_ONLY_16_SCENES = 8,
-	SCENES_ONLY_25_SCENES = 9,
-};
 
 class OBSProjector : public OBSQTDisplay {
 	Q_OBJECT
@@ -45,29 +33,10 @@ private:
 	bool isAlwaysOnTopOverridden = false;
 	int savedMonitor = -1;
 	ProjectorType type = ProjectorType::Source;
-	std::vector<OBSWeakSource> multiviewScenes;
-	std::vector<OBSSource> multiviewLabels;
-	gs_vertbuffer_t *actionSafeMargin = nullptr;
-	gs_vertbuffer_t *graphicsSafeMargin = nullptr;
-	gs_vertbuffer_t *fourByThreeSafeMargin = nullptr;
-	gs_vertbuffer_t *leftLine = nullptr;
-	gs_vertbuffer_t *topLine = nullptr;
-	gs_vertbuffer_t *rightLine = nullptr;
-	// Multiview position helpers
-	float thickness = 4;
-	float offset, thicknessx2 = thickness * 2, pvwprgCX, pvwprgCY, sourceX,
-		      sourceY, labelX, labelY, scenesCX, scenesCY, ppiCX, ppiCY,
-		      siX, siY, siCX, siCY, ppiScaleX, ppiScaleY, siScaleX,
-		      siScaleY, fw, fh, ratio;
+
+	Multiview *multiview = nullptr;
 
 	bool ready = false;
-
-	// argb colors
-	static const uint32_t outerColor = 0xFFD0D0D0;
-	static const uint32_t labelColor = 0xD91F1F1F;
-	static const uint32_t backgroundColor = 0xFF000000;
-	static const uint32_t previewColor = 0xFF00D000;
-	static const uint32_t programColor = 0xFFD00000;
 
 	void UpdateMultiview();
 	void UpdateProjectorTitle(QString name);
